Extract components controllers out of page controllers
