+++ app/static/js/controller/components/MessageController.js
@@ -1,0 +1,13 @@
+export class MessageController {
+
+    constructor(message) {
+        this.message = message;
+    }
+
+    init() {
+        this.message.close_button.addEventListener('click', () => {
+			this.message.close();
+		});
+    }
+
+}
